refactor(plugin): extract acceleration_structure.rs (#1079)
Moves the VulkanAccelerationStructureMethodsVTable callbacks (Linux +
non-Linux variants of the `label` slot) +
HOST_VULKAN_ACCELERATION_STRUCTURE_METHODS_VTABLE + accessor +
acceleration_structure_methods_vtable_null_tests out of
host_services/mod.rs into a sibling submodule. mod.rs re-exports
HOST_VULKAN_ACCELERATION_STRUCTURE_METHODS_VTABLE and
host_vulkan_acceleration_structure_methods_vtable.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
